requirements.txt
streamlit
langchain
python-dotenv
langchain_community